Renamed static folder to components
