+++ client/app/components/logo_upload.jsx
