+++ client/app/components/simple_article.jsx
