+++ client/app/index.jsx
@@ -2,17 +2,17 @@ import React from 'react';
 import {render} from 'react-dom';
 import { Router, Route, IndexRoute, IndexRedirect, hashHistory } from 'react-router';
 
-import App from './static/app.jsx';
-import Home from './static/home.jsx';
-import Login from './static/login.jsx';
-import Article from './static/article.jsx';
-import NewArticle from './static/new.jsx';
-import EditArticle from './static/edit.jsx';
-import ArticleHistory from './static/history.jsx';
-import Admin from './static/admin.jsx';
-import Setup from './static/setup.jsx';
-import EditTopic from './static/edit_topics.jsx';
-import EditUser from './static/edit_users.jsx';
+import App from './components/app.jsx';
+import Home from './components/home.jsx';
+import Login from './components/login.jsx';
+import Article from './components/article.jsx';
+import NewArticle from './components/new.jsx';
+import EditArticle from './components/edit.jsx';
+import ArticleHistory from './components/history.jsx';
+import Admin from './components/admin.jsx';
+import Setup from './components/setup.jsx';
+import EditTopic from './components/edit_topics.jsx';
+import EditUser from './components/edit_users.jsx';
 
 render(
   <Router history={hashHistory}>
